+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/Java8Validator.java
@@ -25,7 +25,7 @@ import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
 import com.github.javaparser.ast.validator.chunks.ModifierValidator;
 
 /**
- * This validator validates according to Java 7 syntax rules.
+ * This validator validates according to Java 8 syntax rules.
  */
 public class Java8Validator extends Java7Validator {
     final Validator modifiersWithoutPrivateInterfaceMethods = new ModifierValidator(true, true, false);