+++ javaparser-core/src/main/java/com/github/javaparser/printer/concretesyntaxmodel/CsmTextBlock.java
@@ -35,9 +35,9 @@ public class CsmTextBlock implements CsmElement {
 
     @Override
     public void prettyPrint(Node node, SourcePrinter printer) {
-        LineEnding lineEnding = node.getLineEndingStyleOrDefault(LineEnding.SYSTEM);
-        printer.print("\"\"\"" + lineEnding.toString());
-        printer.print(property.getValueAsStringAttribute(node));
+        // Note that values within TextBlocks ALWAYS have the \n line ending, per https://openjdk.java.net/jeps/378#1--Line-terminators
+        printer.print("\"\"\"\n");
+        printer.print(property.getValueAsStringAttribute(node)); // TODO: Confirm if we need to force this to use {@code \n} separators
         printer.print("\"\"\"");
     }